pyproject.toml
[tool.poetry]
name = "Novel"
version = "0.1.0"
description = ""
authors = ["hedgehog <42014197+zutomayo147@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.8"
# django = "3.2.16"
django = "3.2.15"
djangorestframework = "3.12.0"
# pygments = "==2.9.0"
djangorestframework-simplejwt = "4.6.0"
djoser = "2.1.0"
django-cors-headers = "3.13.0"
django-filter = "22.1"
# drf-yasg = "1.21.4"
rules = "3.3"
werkzeug = "2.2.2"
django-security = "0.12.0"
# django-storages = "1.13.1"
# drf-social-auth = "0.1.0"
# django-rest-auth = "0.9.5"
# django-allauth = "0.51.0"
Pillow = "9.3.0"
django-environ = "0.9.0"
# pygraphviz = "1.10"
# PyYAML = "6.0"
# django-model-utils = "^4.2.0"
# django-currentuser = "^0.5.3"
# celery = "^5.2.7"
# dotenv = "^0.0.5"
# pandas = "^1.5.0"
# guardian = "^0.2.1"
# dj-database-url = "^1.0.0"
# channels = "^3.0.5"
# nltk = "^3.7"
# gensim = "^4.2.0"
# sklearn = "^0.0"
# django-model-utils = "4.2.0"
# django-maintenance-mode = "0.16.3"
# django-crispy-forms = "1.14.0"
# crispy-bootstrap5 = "0.7"
# drf-nested-routers = "0.93.4"
# drf-writable-nested = "0.7.0"
# django-simple-history = "3.1.1"
gitpython = "3.1.29"
debugpy = "^1.6.3"
dj-database-url = "^1.0.0"
dj-static = "^0.0.6"
python-decouple = "^3.6"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.2.0"

[tool.poetry.group.dev.dependencies]
# factory-boy = "3.2.1"
Faker = "15.1.1"
django-dbbackup = "4.0.2"
django-extensions = "3.2.1"
drf-spectacular = "0.24.2"
drf-typed = "0.2.0"
django-stubs = "1.12.0"
django-debug-toolbar = "3.5.0"
djangorestframework-stubs = "1.7.0"
djangorestframework-types = "0.8.0"
dr-scaffold = "2.1.2"
django-seed = "0.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"